agent: extract name/authorizationToken from config file
The teamcity agent writes its name and authorizationToken back into the
agent configuration file. Chef needs to extracts this information to avoid
overwritting of this information.
Name and authorizationToken are saved as node attributes. If name and
authorizationToken are already set (previous chef run or by administrator)
this information will be used (no extraction).