redesign setting of attribute defaults (work also on other agents names)
A new Teamcity::Agent class is a wrapper around a Mash of agent attributes.
It is used to set the default attribute values for the default agent and other
agents (defined by user) in the same manner without code dublication.
Dependencies between agent attributes are now resolved inside the recipe by
Teamcity::Agent, too.
So attributes can be changed and attributes that depends on this changed value
changes also.
The base attribute is removed because there is not difference between base
and system directory.