release version 0.2.0
The cookbook has been extended to support installation + configuration of
teamcity build agents. All needed features are implemented.

The cookbook is now ready for a wider testing. Automatically testing with
test-kitchen is planed but I currently not know a way to test the cookbook
without a teamcity server and there is not teamcity::server recipe at the
moment.